modified return format of toString function
What changes you have made?

modified return format of toString function from empty string to
points of Line.

Why this change is necessary?

to know what is the data of the object.

What is the impact of the change?

we can know the what is the data of the object.